tectonic/xetex-math.c: break out math routines into this file
This is just a straight cut-and-paste, but hopefully this segregation will
make it possible to make a bunch of variables and functions static.
